refactor to make easier to test
