+++ tests/utils.test.ts
@@ -1,4 +1,4 @@
-import { getDate } from "../app/covid"
+import { getDate } from "../app/utils"
 
 test('adds 1 week to start date', () => {
   expect(getDate(1)).toBe("16-Mar-2020");
